refactor(merchandises) merchandise 리스트 만들어주는 메서드 위치 이동
